feat(m4): @log_inbound + @log_outbound + audit factory
audit/factory.py — settings-driven dispatcher build via the standard provider chain (entry points → builtins → fail). get_dispatcher caches; set_dispatcher / reset_dispatcher for tests. audit/decorators.py — @log_inbound and @log_outbound for sync + async callables. Captures timing, outcome, error class + code, and seeds request_id / correlation_id from ContextVars. Optional payload_factory extracts the audit payload from the call's args/kwargs; raises are swallowed so audit can never break a request.

+++ src/resilience_kit/audit/dispatch.py
@@ -174,8 +174,8 @@ class FireAndForgetDispatcher:
             _BACKEND_BACKOFF_BASE_S * (2 ** (attempt - 1)),
             _BACKEND_BACKOFF_CAP_S,
         )
-        jitter: float = 0.5 + random.random()  # noqa: S311 — backoff jitter, not crypto
-        return base * jitter
+        jitter: float = 0.5 + float(random.random())  # noqa: S311 — backoff jitter, not crypto
+        return float(base * jitter)
 
 
 class InlineDispatcher:
